Power Point Presentation
It showcases the previous sales over time and the stores that has generated the highest revenue and the stores that generated less. Also, made the predictions in Tableau showcasing the perks and downs that are going to be in next few weeks of the sales along with the confidence intervals. ARIMA model is also used in forecasting the sales which clearly shows the predictions.